requirements.txt
nibabel
matplotlib
nilearn
scikit-learn
torcheval
torchmetrics
progress
lolcat
monai